Module GettextStub

Concrete implementation based on native gettext library.

see http://www.gnu.org/software/gettext/gettext.html/

Gettext library

author

Sylvain Le Gall

Concrete implementations

module Native : GettextTypes.REALIZE_TYPE

Native implementation of gettext. All translation are bound to C library call. Still use check_format, to ensure that strings follow printf format.

module Preload : GettextTypes.REALIZE_TYPE

Native implementation of gettext. Use the Native module, but use informations provided to preload all textdomain translation. The preload is made by trying to translate the string "", which is mandatory in MO file. This is not the default behavior of gettext. Use this module if you know that it is better to preload all string. Don't use this module if you think you will only have a few strings to translate.