Module GettextStub.Preload

Native implementation of gettext. Use the Native module, but use informations provided to preload all textdomain translation. The preload is made by trying to translate the string "", which is mandatory in MO file. This is not the default behavior of gettext. Use this module if you know that it is better to preload all string. Don't use this module if you think you will only have a few strings to translate.

val realize : GettextTypes.t -> GettextTypes.t'