Module GettextStubCompat

Low level interface to gettext C library.

author

Sylvain Le Gall

type lc =

| LC_CTYPE

| LC_NUMERIC

| LC_TIME

| LC_COLLATE

| LC_MONETARY

| LC_MESSAGES

| LC_ALL

val setlocale : lc -> string -> string option

Set the current localization for the category

val gettext : string -> string

Look up MSGID in the current default message catalog for the current * LC_MESSAGES locale. If not found, returns MSGID itself (the default text).

val dgettext : string -> string -> string

Look up MSGID in the DOMAINNAME message catalog for the current LC_MESSAGES * locale.

val dcgettext : string -> string -> lc -> string

Look up MSGID in the DOMAINNAME message catalog for the current CATEGORY * locale.

val ngettext : string -> string -> int -> string

Similar to `gettext' but select the plural form corresponding to the number * N.

val dngettext : string -> string -> string -> int -> string

Similar to `dgettext' but select the plural form corresponding to the number * N.

val dcngettext : string -> string -> string -> int -> lc -> string

Similar to `dcgettext' but select the plural form corresponding to the * number N.

val textdomain : string -> string option

Set the current default message catalog to DOMAINNAME.If DOMAINNAME is "", * reset to the default of "messages".

val get_textdomain : unit -> string option

Get the current default message catalog to DOMAINNAME.

val bindtextdomain : string -> string -> string option

Specify that the DOMAINNAME message catalog will be foundin DIRNAME rather * than in the system locale data base.

val bind_textdomain_codeset : string -> string -> string option

Specify the character encoding in which the messages from theDOMAINNAME * message catalog will be returned.